frontend/frontend/requirements.txt
google-generativeai
pdfplumber
nltk
pandas
python-docx
transformers
language_tool_python
spacy
numpy
scikit-learn
textblob
sentence-transformers
pdf2image
Pillow  
